+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Version 0.4.0.dev0
 
 - Add ability to create groups of options with separate panels
+- Show positional arguments in their own panel by default
+- Add config `GROUP_ARGUMENTS_OPTIONS` option to group with options
+- Add support for printing errors nicely
 
 ## Version 0.3.0 (2022-02-13)